+++ Chat/src/Messages/ServerAcceptMessage.java
@@ -3,7 +3,7 @@ package Messages;
 
 public class ServerAcceptMessage extends Message {
 
-	public ServerAcceptMessage(int _op,long _length,long _reserved,String _options,byte[] body){
+	public ServerAcceptMessage(int _op,long _length,long _reserved,String _options,int[] body){
 		super(_op,_length,_reserved,_options);
 
 		if(op!=4){
